feat: header content 분리
