+++ apps/web/components/header/categorySection.tsx
@@ -1,0 +1,61 @@
+'use client';
+
+import { CATEGORY_OPTIONS, CATEGORY_NAME } from 'constants/category';
+import { CategoryOptionEng, CategoryNameEng } from 'types/category';
+import Link from 'next/link';
+import { cn } from '@lococo/utils';
+import { OptionBar } from './categoryOptionBar';
+import { useHeaderAction } from './use-header-action';
+
+export function CategorySection() {
+  const categories = Object.entries(CATEGORY_NAME).map(([key, name]) => ({
+    key: key as CategoryNameEng,
+    name,
+  }));
+  const {
+    selectedCategory,
+    selectedOption,
+    handleSelectCategory,
+    handleSelectOption,
+    handleMouseLeaveCategory,
+  } = useHeaderAction();
+  const options = selectedCategory
+    ? (Object.keys(CATEGORY_OPTIONS[selectedCategory]) as CategoryOptionEng[])
+    : [];
+
+  return (
+    <div
+      className="flex h-[6rem] items-center"
+      onMouseLeave={handleMouseLeaveCategory}
+    >
+      {categories.map(({ key, name }) => {
+        const isActive = key === selectedCategory;
+        return (
+          <div key={key}>
+            <Link
+              href={`/search?middleCategory=${key}&searchType=PRODUCT`}
+              key={key}
+              className={cn(
+                'jp-title2 flex h-[6rem] cursor-pointer items-center whitespace-nowrap px-[3.2rem] pb-[1rem] pt-[1rem] font-bold',
+                isActive ? 'text-pink-500' : 'text-gray-800'
+              )}
+              onMouseEnter={() => handleSelectCategory(key)}
+              onClick={() => handleSelectCategory(key)}
+            >
+              {name}
+            </Link>
+            {selectedCategory && (
+              <OptionBar
+                options={options}
+                selectedOption={selectedOption}
+                selectedCategoryKey={selectedCategory}
+                handleSelectOption={handleSelectOption}
+                handleMouseLeaveCategory={handleMouseLeaveCategory}
+              />
+            )}
+          </div>
+        );
+      })}
+    </div>
+  );
+}
